Move collision event logic into new `ContactReportingPlugin`
Also added `PhysicsStepSet::ReportContacts`
