+++ src/lib.rs
@@ -517,8 +517,9 @@ pub enum PhysicsSet {
 ///     3. Solve positional and angular constraints
 ///     4. Update velocities
 ///     5. Solve velocity constraints (dynamic friction and restitution)
-/// 3. Sleeping
-/// 4. Spatial queries
+/// 3. Report contacts (send collision events)
+/// 4. Sleeping
+/// 5. Spatial queries
 #[derive(SystemSet, Clone, Copy, Debug, PartialEq, Eq, Hash)]
 pub enum PhysicsStepSet {
     /// Responsible for collecting pairs of potentially colliding entities into [`BroadCollisionPairs`] using
@@ -530,6 +531,10 @@ pub enum PhysicsStepSet {
     ///
     /// See [`SubstepSet`] and [`SubstepSchedule`].
     Substeps,
+    /// Responsible for sending collision events and updating [`CollidingEntities`].
+    ///
+    /// See [`ContactReportingPlugin`].
+    ReportContacts,
     /// Responsible for controlling when bodies should be deactivated and marked as [`Sleeping`].
     ///
     /// See [`SleepingPlugin`].
