+++ src/plugins/narrow_phase/mod.rs
@@ -1,11 +1,10 @@
-//! Computes contacts between entities and sends collision events.
+//! Computes contacts between entities.
 //!
 //! See [`NarrowPhasePlugin`].
 
 mod contact_data;
 pub mod contact_query;
 
-use bevy::utils::HashSet;
 pub use contact_data::*;
 pub use contact_query::*;
 
@@ -13,24 +12,17 @@ use crate::prelude::*;
 #[cfg(feature = "parallel")]
 use bevy::tasks::{ComputeTaskPool, ParallelSlice};
 
-/// Computes contacts between entities and sends collision events.
+/// Computes contacts between entities.
 ///
 /// Collisions are only checked between entities contained in [`BroadCollisionPairs`],
 /// which is handled by the [`BroadPhasePlugin`].
 ///
-/// The following collision events are sent each frame:
-///
-/// - [`Collision`]
-/// - [`CollisionStarted`]
-/// - [`CollisionEnded`]
+/// The results of the narrow phase are added into [`Collisions`].
 pub struct NarrowPhasePlugin;
 
 impl Plugin for NarrowPhasePlugin {
     fn build(&self, app: &mut App) {
-        app.add_event::<Collision>()
-            .add_event::<CollisionStarted>()
-            .add_event::<CollisionEnded>()
-            .init_resource::<NarrowPhaseConfig>()
+        app.init_resource::<NarrowPhaseConfig>()
             .init_resource::<Collisions>()
             .register_type::<NarrowPhaseConfig>();
 
@@ -50,10 +42,12 @@ impl Plugin for NarrowPhasePlugin {
                 })
                 .after(PhysicsStepSet::BroadPhase)
                 .before(PhysicsStepSet::Substeps),
-                // Send collision events
-                send_collision_events
-                    .after(PhysicsStepSet::Sleeping)
-                    .before(PhysicsStepSet::SpatialQuery),
+                // Remove ended collisions
+                (|mut collisions: ResMut<Collisions>| {
+                    collisions.retain(|contacts| contacts.during_current_frame)
+                })
+                .after(PhysicsStepSet::ReportContacts)
+                .before(PhysicsStepSet::Sleeping),
             )
                 .chain(),
         );
@@ -102,18 +96,6 @@ impl Default for NarrowPhaseConfig {
         }
     }
 }
-
-/// A [collision event](Collider#collision-events) that is sent for each contact pair during the narrow phase.
-#[derive(Event, Clone, Debug, PartialEq)]
-pub struct Collision(pub Contacts);
-
-/// A [collision event](Collider#collision-events) that is sent when two entities start colliding.
-#[derive(Event, Clone, Debug, PartialEq)]
-pub struct CollisionStarted(pub Entity, pub Entity);
-
-/// A [collision event](Collider#collision-events) that is sent when two entities stop colliding.
-#[derive(Event, Clone, Debug, PartialEq)]
-pub struct CollisionEnded(pub Entity, pub Entity);
 
 #[allow(clippy::too_many_arguments)]
 #[allow(clippy::type_complexity)]
@@ -293,74 +275,4 @@ fn reset_substep_collision_states(mut collisions: ResMut<Collisions>) {
     for contacts in collisions.get_internal_mut().values_mut() {
         contacts.during_current_substep = false;
     }
-}
-
-/// Sends collision events and updates [`CollidingEntities`].
-fn send_collision_events(
-    sleeping: Query<(Ref<Position>, Ref<Rotation>)>,
-    mut colliders: Query<&mut CollidingEntities>,
-    mut collisions: ResMut<Collisions>,
-    mut collision_ev_writer: EventWriter<Collision>,
-    mut collision_started_ev_writer: EventWriter<CollisionStarted>,
-    mut collision_ended_ev_writer: EventWriter<CollisionEnded>,
-) {
-    let mut ended_collisions = HashSet::<(Entity, Entity)>::new();
-
-    for ((entity1, entity2), contacts) in collisions.get_internal_mut().iter_mut() {
-        // Collision ended
-        if !contacts.during_current_frame {
-            // Keep the collision active if the bodies were colliding during the previous frame
-            // but neither body has moved, for example when they are sleeping.
-            if let Ok([(pos1, rot1), (pos2, rot2)]) = sleeping.get_many([*entity1, *entity2]) {
-                if !pos1.is_changed()
-                    && !rot1.is_changed()
-                    && !pos2.is_changed()
-                    && !rot2.is_changed()
-                {
-                    contacts.during_current_frame = true;
-                    continue;
-                }
-            }
-
-            ended_collisions.insert((*entity1, *entity2));
-
-            if let Ok(mut colliding_entities1) = colliders.get_mut(*entity1) {
-                colliding_entities1.remove(entity2);
-            }
-            if let Ok(mut colliding_entities2) = colliders.get_mut(*entity2) {
-                colliding_entities2.remove(entity1);
-            }
-
-            continue;
-        }
-
-        collision_ev_writer.send(Collision(contacts.clone()));
-
-        // Collision started
-        if contacts.during_current_frame && !contacts.during_previous_frame {
-            collision_started_ev_writer.send(CollisionStarted(*entity1, *entity2));
-            contacts.during_previous_frame = true;
-
-            if let Ok(mut colliding_entities1) = colliders.get_mut(*entity1) {
-                colliding_entities1.insert(*entity2);
-            } else {
-                ended_collisions.insert((*entity1, *entity2));
-            }
-            if let Ok(mut colliding_entities2) = colliders.get_mut(*entity2) {
-                colliding_entities2.insert(*entity1);
-            } else {
-                ended_collisions.insert((*entity1, *entity2));
-            }
-        }
-    }
-
-    // Send CollisionEnded events
-    collision_ended_ev_writer.send_batch(
-        ended_collisions
-            .iter()
-            .map(|(entity1, entity2)| CollisionEnded(*entity1, *entity2)),
-    );
-
-    // Clear collisions at the end of each frame to avoid unnecessary iteration and memory usage
-    collisions.retain(|contacts| !ended_collisions.contains(&(contacts.entity1, contacts.entity2)));
 }
