+++ src/plugins/setup.rs
@@ -144,6 +144,7 @@ impl Plugin for PhysicsSetupPlugin {
             (
                 PhysicsStepSet::BroadPhase,
                 PhysicsStepSet::Substeps,
+                PhysicsStepSet::ReportContacts,
                 PhysicsStepSet::Sleeping,
                 PhysicsStepSet::SpatialQuery,
             )
